// src/types/attestation.type.ts
export type attestationType = {
    name: string, // Name of the project
    organizationType: string, // Type of organization
    date: string, // Date of the project
    region: string, // Region of the project
    country: string, // Country of the project
    description: string, // Description of the project
    website: string, // Website of the project
    energyCategory: string, // Energy category of the project
    subCategory: string, // Subcategory of the project
    blockchain: string, // Blockchain of the project
    ods: string[], // ODS of the project
    latitude: string, // Latitude of the project
    longitude: string, // Longitude of the project
    wallet: string // Wallet of the project
};